[Fix] Fix claim command auto complete
